requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysqlclient
pydantic
pydantic-settings
passlib[bcrypt]
PyJWT
python-dotenv
